requirements.txt
python-dotenv
openai
langchain-google-genai
langchain
langchain-openai
google-genai

# PDF Parser dependencies
pdfplumber==0.11.4
Pillow==11.0.0
pydantic==2.10.
pytesseract==0.3.13
numpy<2
pandas==2.2.3
pdf2image==1.17.0